% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_updates.R
\name{check_updates}
\alias{check_updates}
\title{Check for OHA package updates}
\usage{
check_updates(name, url = "https://github.com/USAID-OHA-SI")
}
\arguments{
\item{name}{package name}

\item{url}{user/organization url, default = "https://github.com/USAID-OHA-SI"}
}
\value{
message if there is a newer package on GH than local
}
\description{
This function compares the local version of the package against what is
stored on GitHub to flag whether or not you have the latest version.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
check_updates("gophr")
check_updates("glamr")
\dontshow{\}) # examplesIf}
}
