% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_updates.R
\name{extract_remote_sha}
\alias{extract_remote_sha}
\title{Extract Remote SHA}
\usage{
extract_remote_sha(name, url = "https://github.com/USAID-OHA-SI")
}
\arguments{
\item{name}{package name}

\item{url}{user/organization url, default = "https://github.com/USAID-OHA-SI/"}
}
\value{
40 character SHA hash vector
}
\description{
Pulls the latest SHA (commit ID) from GitHub
}
\keyword{internal}
