% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gagglr-package.R
\docType{package}
\name{gagglr-package}
\alias{gagglr}
\alias{gagglr-package}
\title{gagglr: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
