% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{oha_install_outdated}
\alias{oha_install_outdated}
\title{Install Outdated or Missing OHA Packages}
\usage{
oha_install_outdated(core_only = FALSE)
}
\arguments{
\item{core_only}{only check/install core OHA package? default = FALSE}
}
\description{
Used to automatically install any missing or outdated OHA packages from
GitHub. Ideally running with just core packages and other packages should
be installed individually using \code{pak::pak("USAID-OHA-SI/[name]")}.
}
\examples{
\dontrun{
oha_install_outdated()
}
}
\keyword{internal}
