% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{oha_outdated}
\alias{oha_outdated}
\title{Identify Outdated OHA packages}
\usage{
oha_outdated(core_only = FALSE)
}
\arguments{
\item{core_only}{only check/install core OHA package? default = FALSE}
}
\value{
df of outdated or missing packages
}
\description{
Identify Outdated OHA packages
}
\keyword{internal}
