% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{oha_packages}
\alias{oha_packages}
\title{List all OHA Packages}
\usage{
oha_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include gagglr in the list?}
}
\description{
List all OHA Packages
}
\examples{
oha_packages()
}
