% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_updates.R
\name{print_update_text}
\alias{print_update_text}
\title{Print info for outdated packages}
\usage{
print_update_text(name, org)
}
\arguments{
\item{name}{package name}

\item{org}{GH organization}
}
\description{
Print info for outdated packages
}
\keyword{internal}
