% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{oha_update}
\alias{oha_update}
\title{Update OHA packages}
\usage{
oha_update()
}
\description{
This will check to see if all your OHA packages are up-to-date, and will
install after an interactive confirmation. The versions may match between
GitHub and your local environment, but may still be behind on recent commits
that do not merit a full version update.
}
\examples{
\dontrun{
oha_update()
}
}
